+++ src/components/cards/utils/constants.ts
@@ -1,45 +1,10 @@
 import { globalColors } from '@/design-system';
-
-enum LearnCategory {
-  Essentials = 'essentials',
-  Safety = 'safety',
-  Rainbow = 'rainbow',
-  BeginnersGuides = 'beginners_guides',
-  Blockchains = 'blockchains',
-  Web3 = 'web3',
-}
-
-type LearnCardKey =
-  | 'get_started'
-  | 'backups'
-  | 'protect_wallet'
-  | 'connect_to_dapp'
-  | 'avoid_scams';
-
-type CardColor =
-  | 'pink'
-  | 'yellow'
-  | 'green'
-  | 'blue'
-  | 'purple'
-  | 'darkGreen'
-  | 'darkBlue';
-
-export interface LearnCardDetails {
-  url: string;
-  category: LearnCategory;
-  emoji: string;
-  key: LearnCardKey;
-}
-
-export interface CardColorway {
-  gradient: string[];
-  shadowColor: string;
-  orbColorLight: string;
-  orbColorDark: string;
-  primaryTextColor: string;
-  secondaryTextColor: string;
-}
+import {
+  CardColor,
+  CardColorway,
+  LearnCardDetails,
+  LearnCategory,
+} from './types';
 
 export const getCardColorways: (
   isDarkMode: boolean
