+++ src/components/cards/utils/types.ts
@@ -1,0 +1,40 @@
+export enum LearnCategory {
+  Essentials = 'essentials',
+  Safety = 'safety',
+  Rainbow = 'rainbow',
+  BeginnersGuides = 'beginners_guides',
+  Blockchains = 'blockchains',
+  Web3 = 'web3',
+}
+
+export type LearnCardKey =
+  | 'get_started'
+  | 'backups'
+  | 'protect_wallet'
+  | 'connect_to_dapp'
+  | 'avoid_scams';
+
+export type CardColor =
+  | 'pink'
+  | 'yellow'
+  | 'green'
+  | 'blue'
+  | 'purple'
+  | 'darkGreen'
+  | 'darkBlue';
+
+export type LearnCardDetails = {
+  url: string;
+  category: LearnCategory;
+  emoji: string;
+  key: LearnCardKey;
+};
+
+export type CardColorway = {
+  gradient: string[];
+  shadowColor: string;
+  orbColorLight: string;
+  orbColorDark: string;
+  primaryTextColor: string;
+  secondaryTextColor: string;
+};
